Extract `StockTradeAdmin`
