+++ stocks/admin/__init__.py
@@ -1,0 +1,4 @@
+from .StockTradeAdmin import StockTradeAdmin
+from .admin import *
+
+__all__ = ["StockTradeAdmin"]
